backend/requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.11.3
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.0
pytweening==1.2.0
requests==2.32.3
rich==13.8.1
urllib3==2.2.3
Werkzeug==3.1.3
